Extract base Builder class to reduce duplication and define shared interface
