+++ README.md
@@ -65,15 +65,16 @@ svg = HeatmapBuilder.generate_calendar(scores_by_date)
 ```ruby
 # Customize linear heatmap appearance
 options = {
-  cell_size: 35,           # Size of each square (default: 20)
-  cell_spacing: 1,         # Space between squares (default: 2)
-  font_size: 20,           # Font size for score text (default: 12)
+  cell_size: 35,           # Size of each square (default: 10)
+  cell_spacing: 2,         # Space between squares (default: 1)
+  font_size: 20,           # Font size for score text (default: 8)
+  border_width: 2,         # Border thickness (default: 1)
   colors: %w[
-    #f0f0f0
-    #c6e48b
-    #7bc96f
-    #239a3b
-    #196127
+    #ebedf0
+    #9be9a8
+    #40c463
+    #30a14e
+    #216e39
   ]
 }
 
@@ -85,9 +86,13 @@ svg = HeatmapBuilder.generate([1, 2, 3, 4, 5, 6, 7], options)
 ```ruby
 # Calendar heatmap options
 calendar_options = {
-  cell_size: 14,
-  start_of_week: :sunday,    # :monday (default) or :sunday
-  show_outside_cells: true   # Show cells outside date range
+  cell_size: 14,                    # Size of each cell (default: 12)
+  start_of_week: :sunday,           # :monday (default) or :sunday
+  show_outside_cells: true,         # Show cells outside date range (default: false)
+  show_month_labels: true,          # Show month names (default: true)
+  show_day_labels: true,            # Show day abbreviations (default: true)
+  day_labels: %w[S M T W T F S],    # Custom day labels (default: S M T W T F S)
+  month_labels: %w[Jan Feb Mar Apr May Jun Jul Aug Sep Oct Nov Dec] # Custom month labels
 }
 
 svg = HeatmapBuilder.generate_calendar(scores_by_date, calendar_options)
